+++ packages/react/src/utils/setErrorResponse.ts
@@ -4,6 +4,7 @@ import type {
   FieldValues,
   FormErrors,
   FormStore,
+  Maybe,
   ResponseData,
 } from '../types';
 import { getFieldArrayStore } from './getFieldArrayStore';
@@ -34,7 +35,7 @@ export function setErrorResponse<
 ): void {
   // Combine errors that were not set for any field or field array into one
   // general form error response message
-  const message = Object.entries(formErrors)
+  const message = Object.entries<Maybe<string>>(formErrors)
     .reduce<string[]>((errors, [name, error]) => {
       if (
         [
@@ -46,7 +47,7 @@ export function setErrorResponse<
             (shouldActive && !fieldOrFieldArray.active.peek())
         )
       ) {
-        errors.push(error);
+        errors.push(error!);
       }
       return errors;
     }, [])