+++ packages/react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `getErrors` method (issue #72)
+
 ## v0.1.1 (May 19, 2023)
 
 - Fix bug when using dynamic field arrays